requirements.txt
Django==2.1.1
django-js-asset==1.1.0
django-mptt==0.9.1
pytz==2018.5
gunicorn
Jinja2==2.10
